+++ app/components/inputs/ImageUpload.tsx
@@ -3,12 +3,14 @@
 import { useState } from "react";
 import Image from "next/image";
 import { TbPhotoPlus } from "react-icons/tb";
+import { useTranslation } from "react-i18next";
 
 interface ImageUploadProps {
   onChange: (value: any) => void;
 }
 
 const ImageUpload: React.FC<ImageUploadProps> = ({ onChange }) => {
+  const { t } = useTranslation();
   const [preview, setPreview] = useState<string | null>(null);
 
   const handleUpload = async (event: any) => {
@@ -27,7 +29,9 @@ const ImageUpload: React.FC<ImageUploadProps> = ({ onChange }) => {
       className="relative cursor-pointer hover:opacity-70 transition border-dashed border-2 p-20 border-neutral-300 flex flex-col justify-center items-center gap-4 text-neutral-600"
     >
       <TbPhotoPlus size={50} />
-      <div className="font-semibold text-lg">Натисни, за да качиш снимка</div>
+      <div className="font-semibold text-lg">
+        {t("Press_to_select_a_photo")}
+      </div>
       {preview && (
         <div className="absolute inset-0 w-full h-full">
           <Image