+++ app/components/listings/ListingMap.tsx
@@ -3,12 +3,15 @@
 import { Circle, GoogleMap, useJsApiLoader } from "@react-google-maps/api";
 import Heading from "../Heading";
 import { Listing } from "pawpal-fe-common/listings-types";
+import { mapsStyle } from "pawpal-fe-common/constants";
+import { useTranslation } from "react-i18next";
 
 interface ListingMapProps {
   listing: Listing;
 }
 
 const ListingMap: React.FC<ListingMapProps> = ({ listing }) => {
+  const { t } = useTranslation();
   const { isLoaded } = useJsApiLoader({
     googleMapsApiKey: "AIzaSyCPASOspif-cElvaiBWxsuLwAHKq9YyKbs",
     libraries: ["places"],
@@ -31,136 +34,7 @@ const ListingMap: React.FC<ListingMapProps> = ({ listing }) => {
           zoomControl: false,
           scrollwheel: false,
           disableDoubleClickZoom: true,
-          styles: [
-            {
-              featureType: "administrative",
-              elementType: "labels.text.fill",
-              stylers: [
-                {
-                  lightness: 50,
-                },
-              ],
-            },
-            {
-              featureType: "administrative.country",
-              elementType: "geometry",
-              stylers: [
-                {
-                  lightness: 50,
-                },
-                {
-                  hue: "#000",
-                },
-              ],
-            },
-            {
-              featureType: "landscape",
-              stylers: [
-                {
-                  hue: "#e6f3d6",
-                },
-                {
-                  saturation: 43.400000000000006,
-                },
-                {
-                  lightness: 37.599999999999994,
-                },
-                {
-                  gamma: 1,
-                },
-              ],
-            },
-            {
-              featureType: "water",
-              elementType: "all",
-              stylers: [
-                {
-                  color: "#eaf6f8",
-                },
-                {
-                  visibility: "on",
-                },
-              ],
-            },
-            {
-              featureType: "water",
-              elementType: "geometry.fill",
-              stylers: [
-                {
-                  color: "#eaf6f8",
-                },
-              ],
-            },
-            {
-              featureType: "road",
-              elementType: "geometry",
-              stylers: [
-                {
-                  lightness: 50,
-                },
-                {
-                  visibility: "on",
-                },
-              ],
-            },
-            {
-              featureType: "poi",
-              elementType: "all",
-              stylers: [
-                {
-                  visibility: "off",
-                },
-              ],
-            },
-            {
-              featureType: "poi.park",
-              elementType: "all",
-              stylers: [
-                {
-                  visibility: "on",
-                },
-              ],
-            },
-            {
-              featureType: "poi.park",
-              elementType: "geometry.fill",
-              stylers: [
-                {
-                  color: "#e6f3d6",
-                },
-                {
-                  visibility: "on",
-                },
-              ],
-            },
-            {
-              featureType: "transit",
-              elementType: "all",
-              stylers: [
-                {
-                  visibility: "on",
-                },
-              ],
-            },
-            {
-              featureType: "transit",
-              elementType: "labels.text",
-              stylers: [
-                {
-                  visibility: "off",
-                },
-              ],
-            },
-            {
-              featureType: "water",
-              elementType: "all",
-              stylers: [
-                {
-                  visibility: "on",
-                },
-              ],
-            },
-          ],
+          styles: mapsStyle,
         }}
       >
         <Circle
@@ -179,8 +53,9 @@ const ListingMap: React.FC<ListingMapProps> = ({ listing }) => {
         />
       </GoogleMap>
       <div className="text-xs">
-        * Локацията на картата е ориентировъчна. Точният адрес ще бъде показан
-        след резервация.
+        {t(
+          "The_location_on_the_map_is_approximate_The_exact_address_will_be_displayed_after_reservation"
+        )}
       </div>
     </div>
   ) : (